moving _compat into _ui_tester_register
