+++ traitsui/testing/tester/_ui_tester_register/qt4/_interaction_helpers.py
@@ -11,7 +11,7 @@
 from pyface.qt import QtCore, QtGui
 from pyface.qt.QtTest import QTest
 
-from traitsui.testing.tester._compat import check_key_compat
+from traitsui.testing.tester._ui_tester_register._compat import check_key_compat
 from traitsui.testing.tester.exceptions import Disabled
 from traitsui.qt4.key_event_to_name import key_map as _KEY_MAP
